refactor(backend): utils/ → search/ レイヤー昇格 (global_search / simd_filter) (#448)
