+++ backend/providers/query_provider.cpp
@@ -10,9 +10,9 @@
 #include "../parsers/copy_block_detector.h"
 #include "../parsers/split_utils.h"
 #include "../parsers/sql_parser.h"
+#include "../search/simd_filter.h"
 #include "../utils/json_utils.h"
 #include "../utils/logger.h"
-#include "../utils/simd_filter.h"
 #include "../utils/sql_validation.h"
 #include "../utils/string_utils.h"
 #include "simdjson.h"
