+++ backend/search/global_search.cpp
