+++ backend/search/global_search.h
